Move win32/mingw32 realpath to simulation runtime
realpath is used for URI lookup during runtime.
