+++ Compiler/runtime/systemimpl.c
@@ -106,16 +106,6 @@ typedef void* iconv_t;
 #define S_IFLNK  0120000  /* symbolic link */
 
 #include <sys/types.h>
-#if defined(_MSC_VER)
-  #include <win32_dirent.h>
-  #if !defined(PATH_MAX)
-  #define PATH_MAX MAX_PATH
-  #endif
-#else
-  #include <dirent.h>
-#endif
-
-char *realpath(const char *path, char resolved_path[PATH_MAX]);
 
 #else
 /* includes/defines specific for LINUX/OS X */
@@ -2451,184 +2441,6 @@ const char* SystemImpl__gettext(const char *msgid)
 #endif
 }
 
-
-#if defined(__MINGW32__) || defined(_MSC_VER)
-
-#ifdef __MINGW32__
-
-char *realpath(const char *path, char resolved_path[PATH_MAX])
-{
-  if (!_fullpath(resolved_path, path, PATH_MAX))
-  {
-    const char *c_tokens[0]={};
-    const char* fmt = "System.realpath failed on %s with errno: %d";
-    char* msg = (char*)omc_alloc_interface.malloc_atomic(strlen(path) + strlen(fmt) + 10);
-    sprintf(msg, fmt, path, errno);
-    c_add_message(NULL,6000,
-      ErrorType_scripting,
-      ErrorLevel_warning,
-      msg,
-      c_tokens,
-      0);
-    resolved_path = (char*)path;
-  }
-  return resolved_path;
-}
-
-#else
-
-/*
-realpath() Win32 implementation, supports non standard glibc extension
-This file has no copyright assigned and is placed in the Public Domain.
-Written by Nach M. S. September 8, 2005
-*/
-
-#include <windows.h>
-#include <stdlib.h>
-#include <limits.h>
-#include <errno.h>
-#include <sys/stat.h>
-
-char *realpath(const char *path, char resolved_path[PATH_MAX])
-{
-  char *return_path = 0;
-
-  if (path) //Else EINVAL
-  {
-    if (resolved_path)
-    {
-      return_path = resolved_path;
-    }
-    else
-    {
-      //Non standard extension that glibc uses
-      return_path = (char*)malloc(PATH_MAX);
-    }
-
-    if (return_path) //Else EINVAL
-    {
-      //This is a Win32 API function similar to what realpath() is supposed to do
-      size_t size = GetFullPathNameA(path, PATH_MAX, return_path, 0);
-
-      //GetFullPathNameA() returns a size larger than buffer if buffer is too small
-      if (size > PATH_MAX)
-      {
-        if (return_path != resolved_path) //Malloc'd buffer - Unstandard extension retry
-        {
-          size_t new_size;
-
-          free(return_path);
-          return_path = (char*)malloc(size);
-
-          if (return_path)
-          {
-            new_size = GetFullPathNameA(path, size, return_path, 0); //Try again
-
-            if (new_size > size) //If it's still too large, we have a problem, don't try again
-            {
-              free(return_path);
-              return_path = 0;
-              errno = ENAMETOOLONG;
-            }
-            else
-            {
-              size = new_size;
-            }
-          }
-          else
-          {
-            //I wasn't sure what to return here, but the standard does say to return EINVAL
-            //if resolved_path is null, and in this case we couldn't malloc large enough buffer
-            errno = EINVAL;
-          }
-        }
-        else //resolved_path buffer isn't big enough
-        {
-          return_path = 0;
-          errno = ENAMETOOLONG;
-        }
-      }
-
-      //GetFullPathNameA() returns 0 if some path resolve problem occured
-      if (!size)
-      {
-        if (return_path != resolved_path) //Malloc'd buffer
-        {
-          free(return_path);
-        }
-
-        return_path = 0;
-
-        //Convert MS errors into standard errors
-        switch (GetLastError())
-        {
-          case ERROR_FILE_NOT_FOUND:
-            errno = ENOENT;
-            break;
-
-          case ERROR_PATH_NOT_FOUND: case ERROR_INVALID_DRIVE:
-            errno = ENOTDIR;
-            break;
-
-          case ERROR_ACCESS_DENIED:
-            errno = EACCES;
-            break;
-
-          default: //Unknown Error
-            errno = EIO;
-            break;
-        }
-      }
-
-      //If we get to here with a valid return_path, we're still doing good
-      if (return_path)
-      {
-        struct stat stat_buffer;
-
-        //Make sure path exists, stat() returns 0 on success
-        if (stat(return_path, &stat_buffer))
-        {
-          if (return_path != resolved_path)
-          {
-            free(return_path);
-          }
-
-          return_path = 0;
-          //stat() will set the correct errno for us
-        }
-        //else we succeeded!
-      }
-    }
-    else
-    {
-      errno = EINVAL;
-    }
-  }
-  else
-  {
-    errno = EINVAL;
-  }
-
-  if (return_path == NULL)
-  {
-    const char* fmt = "System.realpath failed on %s with errno: %d";
-    char* msg = (char*)malloc(strlen(path) + strlen(fmt) + 10);
-    sprintf(msg, fmt, path, errno);
-    c_add_message(NULL,6000,
-      ErrorType_scripting,
-      ErrorLevel_warning,
-      msg,
-      NULL,
-      0);
-    resolved_path = (char*)path;
-    return_path = (char*)path;
-  }
-
-  return return_path;
-}
-#endif /* mingw */
-
-#endif /* mingw and msvc */
 
 int System_getTerminalWidth(void)
 {
